+++ hardware/power-libperfmgr/Android.bp
@@ -12,12 +12,30 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
+cc_library {
+    name: "libdisppower-msm8998",
+    proprietary: true,
+    srcs: ["disp-power/InteractionHandler.cpp"],
+    shared_libs: [
+        "libbase",
+        "libcutils",
+        "liblog",
+        "libperfmgr",
+        "libutils",
+    ],
+}
+
 cc_binary {
     name: "android.hardware.power@1.3-service.yoshino",
     relative_install_path: "hw",
-    init_rc: ["android.hardware.power@1.3-service.yoshino.rc"],
-    vintf_fragments: ["android.hardware.power@1.3-service.yoshino.xml"],
-    srcs: ["service.cpp", "Power.cpp", "InteractionHandler.cpp", "power-helper.c"],
+    vintf_fragments: ["hidl/android.hardware.power@1.3-service.yoshino.xml"],
+    init_rc: ["hidl/android.hardware.power@1.3-service.yoshino.rc"],
+    srcs: [
+        "hidl/service.cpp",
+        "hidl/Power.cpp",
+        "hidl/power-helper.c",
+    ],
     cflags: [
         "-Wall",
         "-Werror",
@@ -32,6 +50,7 @@ cc_binary {
         "android.hardware.power@1.1",
         "android.hardware.power@1.2",
         "android.hardware.power@1.3",
+        "libdisppower-msm8998",
         "libperfmgr",
     ],
     proprietary: true,
